pyproject.toml
[tool.poetry]
name = "intelisort"
version = "0.1.0"
description = ""
authors = ["HRNPH <hrnph@protonmail.com>"]
license = "TheCroissant"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
prisma = "^0.12.0"
gunicorn = "^21.2.0"
loguru = "^0.7.2"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
pydantic = "^2.6.1"
python-multipart = "^0.0.9"
modal = "^0.57.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
